model_definitions/pima_h2o_xgb/model_modules/requirements.txt
teradataml==17.20.0.3
aoa==7.0.1
pandas==1.1.5
matplotlib==3.5.2
scikit-learn==0.24.2
h2o==3.40.0.4
install-jdk==1.0.4
